requirements.txt
aiofiles==0.7.0
aiosqlite==0.17.0
amqp==5.0.6
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2020.11.8
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.0.post0
cycler==0.10.0
dashboard==0.0.6
databases==0.4.3
dj-database-url==0.5.0
Django==3.2.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
h11==0.12.0
idna==2.10
Jinja2==3.0.1
kaleido==0.2.1
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
orm==0.2.0.dev1
panda==0.3.1
pandas==1.2.5
Pillow==8.2.0
plotly==5.0.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
requests==2.25.0
retrying==1.3.3
rq==1.9.0
scipy==1.6.3
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.14.2
tenacity==7.0.0
typesystem==0.3.0.dev0
typing-extensions==3.10.0.0
urllib3==1.26.2
uvicorn==0.13.4
vine==5.0.0
wallstreet==0.3
wcwidth==0.2.5
whitenoise==5.3.0
